Revert "Practice commit 44~46"
