+++ package.json
@@ -1,5 +1,5 @@
 {
-  "main": "src/screens/index.js",
+  "main": "index.js",
   "scripts": {
     "android": "react-native run-android",
     "ios": "react-native run-ios",
