--- src/api/CityListApi.js
@@ -1,6 +1,0 @@
-class CityListApi {
-    fetchAvailableCities = () => fetch('https://raw.githubusercontent.com/example0312/weather-crawler/master/availableCityNames')
-        .then(response => response.json());
-}
-
-export default new CityListApi();
